+++ package.json
@@ -7,6 +7,7 @@
     "graphql": "^16.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-router-dom": "^6.8.0",
     "react-scripts": "5.0.1",
     "web-vitals": "2.1.4"
   },
