+++ src/pages/main-page.tsx
@@ -1,0 +1,3 @@
+export function MainPage() {
+  return <div></div>;
+}
